add new customhook
